+++ hexa/core/models/behaviors.py
@@ -1,12 +1,17 @@
-class WithStatus:
+import enum
+
+
+class Status(enum.Enum):
     SUCCESS = "SUCCESS"
     PENDING = "PENDING"
     RUNNING = "RUNNING"
     ERROR = "ERROR"
     UNKNOWN = "UNKNOWN"
 
+
+class WithStatus:
     @property
-    def status(self):
+    def status(self) -> Status:
         raise NotImplementedError(
             "Classes having the WithStatus behavior should implement status()"
         )