+++ hexa/core/tests/test_views.py
@@ -1,6 +1,7 @@
 from django import test
 from django.urls import reverse
 
+from hexa.core.datagrids import ActivityGrid
 from hexa.user_management.models import User
 
 
@@ -16,6 +17,13 @@ class CoreTest(test.TestCase):
         response = self.client.get(reverse("core:index"))
 
         self.assertEqual(response.status_code, 200)
+
+    def test_dashboard_200(self):
+        self.client.force_login(self.USER_REGULAR)
+        response = self.client.get(reverse("core:dashboard"))
+
+        self.assertEqual(response.status_code, 200)
+        self.assertIsInstance(response.context["last_activity_grid"], ActivityGrid)
 
     def test_ready_200(self):
         response = self.client.get(reverse("core:ready"))