+++ hexa/pipelines/urls.py
@@ -6,7 +6,6 @@ app_name = "pipelines"
 
 urlpatterns = [
     path("", views.index, name="index"),
-    path("refresh", views.index_refresh, name="index_refresh"),
     path(
         "sync/<int:environment_contenttype_id>/<uuid:environment_id>",
         views.environment_sync,