+++ hexa/plugins/connector_airflow/apps.py
@@ -7,6 +7,10 @@ class ConnectorAirflowConfig(ConnectorAppConfig):
 
     verbose_name = "Airflow Connector"
 
+    LAST_ACTIVITIES = (
+        "hexa.plugins.connector_airflow.last_activities.get_last_activities"
+    )
+
     @property
     def route_prefix(self):
         return "airflow"