+++ hexa/ui/datagrid.py
@@ -7,10 +7,11 @@ from django.db import models
 from django.http import HttpRequest
 from django.template import loader
 from django.utils import timezone
+from django.utils.safestring import mark_safe
 from django.utils.timesince import timesince
 from django.utils.translation import ugettext_lazy as _
 
-from hexa.core.models import WithStatus
+from hexa.core.models.behaviors import Status
 from hexa.ui.utils import get_item_value
 
 DjangoModel = typing.TypeVar("DjangoModel", bound=models.Model)
@@ -210,6 +211,8 @@ class LeadingColumn(Column):
         image_src=None,
         icon=None,
         translate=True,
+        bold=True,
+        mark_safe=False,
         **kwargs,
     ):
         super().__init__(**kwargs)
@@ -219,6 +222,8 @@ class LeadingColumn(Column):
         self.image_src = image_src
         self.icon = icon
         self.translate = translate
+        self.bold = bold
+        self.mark_safe = mark_safe
 
     @property
     def template(self):
@@ -226,10 +231,13 @@ class LeadingColumn(Column):
 
     def context(self, model: DjangoModel, grid: Datagrid):
         text_value = self.get_value(model, self.text, container=grid)
+        if self.mark_safe:
+            text_value = mark_safe(text_value)
         data = {
             "text": text_value,
             "single": self.secondary_text is None,
             "translate": self.translate,
+            "bold": self.bold,
         }
         if (
             self.detail_url is None
@@ -245,6 +253,8 @@ class LeadingColumn(Column):
             secondary_text_value = self.get_value(
                 model, self.secondary_text, container=grid
             )
+            if self.mark_safe:
+                secondary_text_value = mark_safe(secondary_text_value)
             data.update(
                 secondary_text=secondary_text_value,
                 empty=text_value is None and secondary_text_value is None,
@@ -399,19 +409,19 @@ class StatusColumn(Column):
         self.value = value
 
     COLOR_MAPPINGS = {
-        WithStatus.SUCCESS: "green",
-        WithStatus.ERROR: "red",
-        WithStatus.RUNNING: "yellow",
-        WithStatus.PENDING: "gray",
-        WithStatus.UNKNOWN: "gray",
+        Status.SUCCESS: "green",
+        Status.ERROR: "red",
+        Status.RUNNING: "yellow",
+        Status.PENDING: "gray",
+        Status.UNKNOWN: "gray",
     }
 
     LABEL_MAPPINGS = {
-        WithStatus.SUCCESS: _("Success"),
-        WithStatus.ERROR: _("Error"),
-        WithStatus.RUNNING: _("Running"),
-        WithStatus.PENDING: _("Pending"),
-        WithStatus.UNKNOWN: _("Unknown"),
+        Status.SUCCESS: _("Success"),
+        Status.ERROR: _("Error"),
+        Status.RUNNING: _("Running"),
+        Status.PENDING: _("Pending"),
+        Status.UNKNOWN: _("Unknown"),
     }
 
     @property